Changes: - Updated code so it uses NonBlocking ConcurrentDictionary

+++ HoneyPot/HoneyPot.Api/HoneyAmount.cs
@@ -2,6 +2,6 @@
 {
     public class HoneyAmount
     {
-        public int TotalAmount { get; set; }
+        public long TotalAmount { get; set; }
     }
 }
